+++ NEWS.md
@@ -1,3 +1,9 @@
+# dplyrUtil 0.0.3
+
+- issues warnings on usage of mapGroups, expandAllInconsistentFactorLevels,
+  expandFactorLevels, and left_joinFactors, because recent dplyr can deal
+  with those issues
+
 # dplyrUtil 0.0.2
 
 allow mutate_cond to work on conditions with NA